Lost Packet improvements and other minor enhancements, including: - Number of lost packets retrieved is now displayed in the stats menu - Lost Packets are smaller and move slightly quicker - Better lighting performance on the packet spotlight - Smoother movement of packets (no more acceleration) - Adjusting camera bounds - Better scaled automation implementation - Better bits/sec calculation for the top menu

+++ Assets/Scripts/Managers/StorageUnitManager.cs
@@ -39,7 +39,7 @@ public class StorageUnitManager : MonoBehaviour {
 		return GameManager.Instance.GameState.StorageCapacity * GameManager.Instance.UpgradeManager.UpgradeState.StorageCapacity;
 	}
 
-	public void AddBits(int bits) {
+	public void AddBits(float bits) {
 		if (bits <= 0) {
 			return;
 		}
